requirements.txt
google-auth-oauthlib==1.0.0
google-api-python-client==2.86.0
python-dateutil==2.8.2
streamlit==1.32.0
pandas==2.2.0 